+++ src/box/constraint_id.c
@@ -37,7 +37,6 @@
 const char *constraint_type_strs[] = {
 	[CONSTRAINT_TYPE_PK]		= "PRIMARY KEY",
 	[CONSTRAINT_TYPE_UNIQUE]	= "UNIQUE",
-	[CONSTRAINT_TYPE_FK]		= "FOREIGN KEY",
 	[CONSTRAINT_TYPE_CK]		= "CHECK",
 };
